Task_7 - Refactor factory - Moved the different type of webdriver setupd to separate creator classes to better follow the factory pattern when creating webdriver
